+++ relay-metrics/src/aggregatorservice.rs
@@ -8,7 +8,7 @@ use relay_system::{
 };
 use serde::{Deserialize, Serialize};
 
-use crate::aggregator::{self, AggregatorConfig, ShiftKey};
+use crate::aggregator::{self, AggregatorConfig, FlushBatching};
 use crate::bucket::Bucket;
 use crate::statsd::{MetricCounters, MetricHistograms, MetricTimers};
 
@@ -86,12 +86,6 @@ pub struct AggregatorServiceConfig {
     /// Defaults to `None`, i.e. no limit.
     pub max_project_key_bucket_bytes: Option<usize>,
 
-    /// Key used to shift the flush time of a bucket.
-    ///
-    /// This prevents flushing all buckets from a bucket interval at the same
-    /// time by computing an offset from the hash of the given key.
-    pub shift_key: ShiftKey,
-
     // TODO(dav1dde): move these config values to a better spot
     /// The approximate maximum number of bytes submitted within one flush cycle.
     ///
@@ -100,11 +94,23 @@ pub struct AggregatorServiceConfig {
     /// adds some additional overhead, this number is approxmate and some safety margin should be
     /// left to hard limits.
     pub max_flush_bytes: usize,
+
     /// The number of logical partitions that can receive flushed buckets.
     ///
     /// If set, buckets are partitioned by (bucket key % flush_partitions), and routed
     /// by setting the header `X-Sentry-Relay-Shard`.
     pub flush_partitions: Option<u64>,
+
+    /// The batching mode for the flushing of the aggregator.
+    ///
+    /// Batching is applied via shifts to the flushing time that is determined when the first bucket
+    /// is inserted. Thanks to the shifts, Relay is able to prevent flushing all buckets from a
+    /// bucket interval at the same time.
+    ///
+    /// For example, the aggregator can choose to shift by the same value all buckets within a given
+    /// partition, effectively allowing all the elements of that partition to be flushed together.
+    #[serde(alias = "shift_key")]
+    pub flush_batching: FlushBatching,
 }
 
 impl Default for AggregatorServiceConfig {
@@ -120,9 +126,9 @@ impl Default for AggregatorServiceConfig {
             max_tag_key_length: 200,
             max_tag_value_length: 200,
             max_project_key_bucket_bytes: None,
-            shift_key: ShiftKey::default(),
             max_flush_bytes: 5_000_000, // 5 MB
             flush_partitions: None,
+            flush_batching: FlushBatching::Project,
         }
     }
 }
@@ -139,7 +145,8 @@ impl From<&AggregatorServiceConfig> for AggregatorConfig {
             max_tag_key_length: value.max_tag_key_length,
             max_tag_value_length: value.max_tag_value_length,
             max_project_key_bucket_bytes: value.max_project_key_bucket_bytes,
-            shift_key: value.shift_key,
+            flush_partitions: value.flush_partitions,
+            flush_batching: value.flush_batching,
         }
     }
 }
@@ -219,6 +226,10 @@ pub struct BucketCountInquiry;
 ///   failed buckets. They will be merged back into the aggregator and flushed at a later time.
 #[derive(Clone, Debug)]
 pub struct FlushBuckets {
+    /// The partition to which the buckets belong.
+    ///
+    /// When set to `Some` it means that partitioning was enabled in the [`Aggregator`].
+    pub partition_key: Option<u64>,
     /// The buckets to be flushed.
     pub buckets: HashMap<ProjectKey, Vec<Bucket>>,
 }
@@ -277,23 +288,25 @@ impl AggregatorService {
     /// If `force` is true, flush all buckets unconditionally and do not attempt to merge back.
     fn try_flush(&mut self) {
         let force_flush = matches!(&self.state, AggregatorState::ShuttingDown);
-        let buckets = self.aggregator.pop_flush_buckets(force_flush);
+        let partitions = self.aggregator.pop_flush_buckets(force_flush);
 
-        if buckets.is_empty() {
+        if partitions.is_empty() {
             return;
         }
 
-        relay_log::trace!("flushing {} projects to receiver", buckets.len());
+        relay_log::trace!("flushing {} partitions to receiver", partitions.len());
 
         let mut total_bucket_count = 0u64;
-        for buckets in buckets.values() {
-            let bucket_count = buckets.len() as u64;
-            total_bucket_count += bucket_count;
+        for buckets_by_project in partitions.values() {
+            for buckets in buckets_by_project.values() {
+                let bucket_count = buckets.len() as u64;
+                total_bucket_count += bucket_count;
 
-            relay_statsd::metric!(
-                histogram(MetricHistograms::BucketsFlushedPerProject) = bucket_count,
-                aggregator = self.aggregator.name(),
-            );
+                relay_statsd::metric!(
+                    histogram(MetricHistograms::BucketsFlushedPerProject) = bucket_count,
+                    aggregator = self.aggregator.name(),
+                );
+            }
         }
 
         relay_statsd::metric!(
@@ -302,7 +315,12 @@ impl AggregatorService {
         );
 
         if let Some(ref receiver) = self.receiver {
-            receiver.send(FlushBuckets { buckets })
+            for (partition_key, buckets_by_project) in partitions {
+                receiver.send(FlushBuckets {
+                    partition_key,
+                    buckets: buckets_by_project,
+                })
+            }
         }
     }