+++ relay-server/src/metrics_extraction/transactions/mod.rs
@@ -13,7 +13,7 @@ use relay_metrics::{Bucket, DurationUnit, FiniteF64};
 use crate::metrics_extraction::generic;
 use crate::metrics_extraction::transactions::types::{
     CommonTag, CommonTags, ExtractMetricsError, LightTransactionTags, TransactionCPRTags,
-    TransactionDurationTags, TransactionMeasurementTags, TransactionMetric, UsageTags,
+    TransactionMeasurementTags, TransactionMetric, UsageTags,
 };
 use crate::metrics_extraction::IntoMetric;
 use crate::statsd::RelayCounters;
@@ -394,20 +394,11 @@ impl TransactionExtractor<'_> {
         // Duration
         let duration = relay_common::time::chrono_to_positive_millis(end - start);
         if let Some(duration) = FiniteF64::new(duration) {
-            let has_profile = if self.config.version >= 3 {
-                false
-            } else {
-                self.has_profile
-            };
-
             metrics.project_metrics.push(
                 TransactionMetric::Duration {
                     unit: DurationUnit::MilliSecond,
                     value: duration,
-                    tags: TransactionDurationTags {
-                        has_profile,
-                        universal_tags: tags.clone(),
-                    },
+                    tags: tags.clone(),
                 }
                 .into_metric(timestamp),
             );
@@ -641,7 +632,6 @@ mod tests {
                     2020-08-21T02:18:20Z,
                 ),
                 exclusive_time: 2000.0,
-                description: "<OrganizationContext>",
                 op: "react.mount",
                 span_id: SpanId(
                     "bd429c44b67a3eb4",
@@ -655,6 +645,7 @@ mod tests {
                 segment_id: ~,
                 is_segment: ~,
                 status: ~,
+                description: "<OrganizationContext>",
                 tags: ~,
                 origin: ~,
                 profile_id: ~,