+++ relay-spans/src/span.rs
@@ -602,7 +602,6 @@ mod tests {
                 1970-01-01T00:02:03Z,
             ),
             exclusive_time: 500.0,
-            description: ~,
             op: "myop",
             span_id: SpanId(
                 "fa90fdead5f74052",
@@ -618,6 +617,7 @@ mod tests {
             ),
             is_segment: ~,
             status: Ok,
+            description: ~,
             tags: ~,
             origin: ~,
             profile_id: EventId(